+++ apps/frontend/components.json
@@ -13,7 +13,7 @@
     "iconLibrary": "lucide",
     "aliases": {
         "components": "@/components",
-        "utils": "@/lib/utils",
+        "utils": "@/lib/utils/shadcn-helper",
         "ui": "@/components/primitives",
         "lib": "@/lib",
         "hooks": "@/hooks"
