+++ apps/frontend/src/components/primitives/field.tsx
@@ -3,7 +3,7 @@
 import { useMemo } from "react"
 import { cva, type VariantProps } from "class-variance-authority"
 
-import { cn } from "@/lib/utils"
+import { cn } from "@/lib/utils/shadcn-helper"
 import { Label } from "@/components/primitives/label"
 import { Separator } from "@/components/primitives/separator"
