+++ apps/frontend/src/components/primitives/navigation-menu.tsx
@@ -3,7 +3,7 @@ import { cva } from "class-variance-authority"
 import { ChevronDownIcon } from "lucide-react"
 import { NavigationMenu as NavigationMenuPrimitive } from "radix-ui"
 
-import { cn } from "@/lib/utils"
+import { cn } from "@/lib/utils/shadcn-helper"
 
 function NavigationMenu({
   className,
