+++ apps/frontend/src/components/primitives/skeleton.tsx
@@ -1,4 +1,4 @@
-import { cn } from "@/lib/utils"
+import { cn } from "@/lib/utils/shadcn-helper"
 
 function Skeleton({ className, ...props }: React.ComponentProps<"div">) {
   return (
